Remove another commit-message-as-a-comment
The fix commented relates to Django 1.7's removing of `mime_type`
(in favor of `content_type`) in `HttpResponse`.

Particularly, the deprecation docs at https://docs.djangoproject.com/en/dev/internals/deprecation/#deprecation-removed-in-1-7
state that

"The mimetype argument to the __init__ methods of HttpResponse,
SimpleTemplateResponse, and TemplateResponse, will be removed.
content_type should be used instead. This also applies to the
render_to_response() shortcut and the sitemap views, index() and
sitemap()."